Added delete conform modal

+++ src/app/components/letter-list/letter-list.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Subscription } from 'rxjs';
 import { LetterService } from 'src/app/services/letter.service';
 import { Letter } from 'src/app/shared/types';
+import { DeleteConfirmDialogComponent } from '../delete-confirm-dialog/delete-confirm-dialog.component';
 
 @Component({
   selector: 'app-letter-list',
@@ -11,7 +13,10 @@ import { Letter } from 'src/app/shared/types';
 export class LetterListComponent implements OnInit, OnDestroy {
   letters: Letter[] = [];
   subscription: Subscription = new Subscription();
-  constructor(private letterService: LetterService) {}
+  constructor(
+    private letterService: LetterService,
+    private dialog: MatDialog
+  ) {}
 
   ngOnInit(): void {
     this.subscription.add(
@@ -27,6 +32,11 @@ export class LetterListComponent implements OnInit, OnDestroy {
 
   onDelete(event: Event, id: string) {
     event.stopPropagation();
-    this.letterService.deleteLetter(id);
+    const dialogRef = this.dialog.open(DeleteConfirmDialogComponent);
+    dialogRef.afterClosed().subscribe((res) => {
+      if (res === true) {
+        this.letterService.deleteLetter(id);
+      }
+    });
   }
 }
